server/requirements.txt
psycopg2===2.9.7
python-dotenv===1.0.0
sqlalchemy===1.4.39